+++ isy-ueberwachung/src/main/java/de/bund/bva/isyfact/ueberwachung/common/konstanten/EreignisSchluessel.java
@@ -21,11 +21,24 @@ package de.bund.bva.isyfact.ueberwachung.common.konstanten;
  */
 public class EreignisSchluessel {
 
+    /** Standard-Ereignis für Debug/Info-Ausgaben ohne eigene Ereignis-ID. */
     public static final String PLUEB00001 = "PLUEB00001";
 
+    /** Pruefung der Pruefroutine nicht erfolgreich.
+     * Wird nicht mehr verwendet. */
+    @Deprecated
     public static final String PLUEB00002 = "PLUEB00002";
 
+    /** Pruefung der Pruefroutine nicht erfolgreich (hat Exception geworfen).
+     * Wird nicht mehr verwendet. */
+    @Deprecated
     public static final String PLUEB00003 = "PLUEB00003";
 
+    /** Selbsttest fehlgeschlagen.
+     * Wird nicht mehr verwendet. */
+    @Deprecated
     public static final String PLUEB00004 = "PLUEB00004";
+
+    /** Loadbalancer wurde abgefragt, aber isAlive-Datei existiert nicht. */
+    public static final String IS_ALIVE_EXISTIERT_NICHT = "PLUEB00005";
 }