+++ isy-ueberwachung/src/main/java/de/bund/bva/isyfact/ueberwachung/service/loadbalancer/LoadbalancerServlet.java
@@ -19,7 +19,6 @@ package de.bund.bva.isyfact.ueberwachung.service.loadbalancer;
 import java.io.File;
 import java.io.IOException;
 
-import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -65,7 +64,8 @@ public class LoadbalancerServlet extends HttpServlet {
 
         String isAliveFileLocation = getInitParameter(PARAM_IS_ALIVE_FILE_LOCATION);
         if (isAliveFileLocation == null) {
-            LOG.debug("Position der IsAliveDatei nicht konfiguriert. Verwende Standard-Einstellung: {}",
+            LOG.info(LogKategorie.JOURNAL, EreignisSchluessel.PLUEB00001,
+                "Position der IsAliveDatei nicht konfiguriert. Verwende Standard-Einstellung: {}",
                 DEFAULT_IS_ALIVE_FILE_LOCATION);
             isAliveFileLocation = DEFAULT_IS_ALIVE_FILE_LOCATION;
         }
@@ -76,16 +76,16 @@ public class LoadbalancerServlet extends HttpServlet {
             isAliveFile.getAbsolutePath());
     }
 
-    /**
+    /** GET-Request bearbeiten. Prüft, ob die IsAlive-Datei vorhanden ist und liefert dann HTTP OK zurück.
+     * Andernfalls wird HTTP FORBIDDEN zurückgeliefert.
+     *
      * @param req
      *            Der HttpServletRequest an das Loadbalancer-Servlet.
      * @param resp
      *            Die Antwort des Loadbalancer-Servlets.
      *
      * @throws IOException
-     *             Wenn die Antwort nicht geschrieben werden kann. GET-Request bearbeiten. Prüft, ob die
-     *             IsAlive-Datei vorhanden ist und liefert dann HTTP  OK zurück. Andernfalls wird
-     *             HTTP FORBIDDEN zurückgeliefert.
+     *             Wenn die Antwort nicht geschrieben werden kann.
      */
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException {
@@ -94,10 +94,10 @@ public class LoadbalancerServlet extends HttpServlet {
             resp.setStatus(HttpServletResponse.SC_OK);
             resp.getWriter().write("<html><body><center>IS ALIVE!</center></body></html>");
         } else {
-            LOG.info(LogKategorie.JOURNAL, EreignisSchluessel.PLUEB00001,
+            LOG.error(
+                    EreignisSchluessel.IS_ALIVE_EXISTIERT_NICHT,
                 "IsAlive-Datei {} existiert nicht, sende HTTP FORBIDDEN.", isAliveFile.getAbsolutePath());
             resp.setStatus(HttpServletResponse.SC_FORBIDDEN);
         }
     }
-
 }