+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 2.2.0
+- `IFS-453`: [isy-ueberwachung] Loglevel für isAlive-Datei-Ereignisse erhöht.
+
 # 2.1.0
 - `IFS-223`: Logging Bridges in products-pom ergänzt: jcl-over-slf4j, slf4j-jcl
 - `IFS-334`: Hinzufügen eines Profils für Jacoco und Anpassung der Surefire Config